+++ CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGELOG
 * @hartman added default print styling ([view](https://github.com/videojs/video.js/pull/3304))
 * @ldayananda updated videojs to not do anything if no src is set ([view](https://github.com/videojs/video.js/pull/3378))
 * @nickygerritsen removed unused tracks when changing sources. Fixes #3000 ([view](https://github.com/videojs/video.js/pull/3002))
+* @vit-koumar updated Flash tech to return Infinity from duration instead of -1 ([view](https://github.com/videojs/video.js/pull/3128))
 
 --------------------